Created IdelPog.Validation.Tests and moved all the assertion lib tests into it
