+++ IdelPog.Tests/Assertions/Pipelines/RepositoryAsserterTest.cs
@@ -3,7 +3,7 @@ using IdelPog.Validation.Assertions;
 using IdelPog.Validation.Assertions.Handlers;
 using IdelPog.Validation.Exceptions;
 
-namespace IdelPogTests.Validation.Pipelines
+namespace IdelPogTests.Assertions.Pipelines
 {
     [TestFixture]
     public class RepositoryAsserterTest
