+++ IdelPog.sln
@@ -13,6 +13,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{1361A0B4-333
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{225E53A1-4F82-4AD5-B50B-70F56F93508E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IdelPog.Validation.Tests", "IdelPog.Validation.Tests\IdelPog.Validation.Tests.csproj", "{800D7FED-8DFC-4EC6-B537-09D210EEBD29}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +33,10 @@ Global
 		{4C55A06A-DCA1-4C47-9F1D-73C8E95C0D6E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4C55A06A-DCA1-4C47-9F1D-73C8E95C0D6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4C55A06A-DCA1-4C47-9F1D-73C8E95C0D6E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{800D7FED-8DFC-4EC6-B537-09D210EEBD29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{800D7FED-8DFC-4EC6-B537-09D210EEBD29}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{800D7FED-8DFC-4EC6-B537-09D210EEBD29}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{800D7FED-8DFC-4EC6-B537-09D210EEBD29}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -39,5 +45,6 @@ Global
 		{21BD9257-C888-476C-9813-B127AB7213B4} = {1361A0B4-3337-4451-9BA2-5EF0EB914EA4}
 		{14CFC328-71E4-4054-9A3C-3EC34F340205} = {225E53A1-4F82-4AD5-B50B-70F56F93508E}
 		{4C55A06A-DCA1-4C47-9F1D-73C8E95C0D6E} = {1361A0B4-3337-4451-9BA2-5EF0EB914EA4}
+		{800D7FED-8DFC-4EC6-B537-09D210EEBD29} = {225E53A1-4F82-4AD5-B50B-70F56F93508E}
 	EndGlobalSection
 EndGlobal
